pyproject.toml
[project]
name = "presentation-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-pptx>=1.0.2",
    "pdfplumber>=0.11.0",
    "gigachat>=0.1.0",
    "python-dotenv>=1.0.0",
]
